clean build with -Wall -Wextra -pedantic

+++ Draw.cc
@@ -127,7 +127,7 @@ auto LevelDraw::drawParticleEffects (const Level& level) -> void
             }
         }
     // Manage existing particles
-    for ( int i = 0; i < m_particles.size(); ++i ) 
+    for ( size_t i = 0; i < m_particles.size(); ++i ) 
         {
         Particle * part = &m_particles[i];
         part->ticks++;
